linting and formatting

+++ app/services/category.py
@@ -1,11 +1,12 @@
 from sqlalchemy.orm import Session
 
 from app.repositories.category import CategoryRepository
-from app.schemas.category import CategoryRead, CategoryCreate, CategoryUpdate
+from app.schemas.category import CategoryCreate, CategoryRead, CategoryUpdate
 
-	
+
 class CategoryNotFoundError(Exception):
     pass
+
 
 class CategoryService:
     """Ключевые операции с задачами, включая бизнес-логику, валидацию и прочее"""
@@ -16,15 +17,21 @@ class CategoryService:
 
     def list_categories(self) -> list[CategoryRead]:
         categories = self.repository.get_all()
-        return [CategoryRead.model_validate(category) for category in categories] # список Pydantic моделей
+        return [
+            CategoryRead.model_validate(category) for category in categories
+        ]  # список Pydantic моделей
 
     def create_category(self, payload: CategoryCreate) -> CategoryRead:
         category = self.repository.create(name=payload.name)
         self.db.commit()
         return CategoryRead.model_validate(category)
 
-    def update_category(self, category_id: str, payload: CategoryUpdate) -> CategoryRead:
+    def update_category(
+        self, category_id: str, payload: CategoryUpdate
+    ) -> CategoryRead:
         category = self.repository.get_by_id(category_id)
+        if category is None:
+            raise CategoryNotFoundError(f"Task {category_id} not found")
 
         if payload.name is not None:
             category.name = payload.name
@@ -34,6 +41,8 @@ class CategoryService:
 
     def delete_category(self, category_id: str) -> None:
         category = self.repository.get_by_id(category_id)
+        if category is None:
+            raise CategoryNotFoundError(f"Task {category_id} not found")
 
         self.repository.delete(category)
         self.db.commit()
